Refactor and update _test impls
