+++ v2client/algod/v2wrappers.go
@@ -3,23 +3,24 @@ package algod
 import (
 	"context"
 	"fmt"
-	"github.com/algorand/go-algorand-sdk/client/algod/models"
 	"github.com/algorand/go-algorand-sdk/types"
+	"github.com/algorand/go-algorand-sdk/v2client/common"
+	"github.com/algorand/go-algorand-sdk/v2client/common/models"
 )
 
 // TODO ejr handling of token
 // TODO ejr received peer feedback to have Block return types.Block not models.Block?
-func (client Client) Shutdown(ctx context.Context, timeout models.ShutdownParams, headers ...*Header) error {
+func (client ClientV2) Shutdown(ctx context.Context, timeout models.ShutdownParams, headers ...*common.Header) error {
 	// TODO EJR how to handle private security token
 	return client.post(ctx, nil, "/shutdown", timeout, headers)
 }
 
-func (client Client) RegisterParticipationKeys(ctx context.Context, account string, params models.RegisterParticipationKeysAccountIdParams, headers ...*Header) error {
+func (client ClientV2) RegisterParticipationKeys(ctx context.Context, account string, params models.RegisterParticipationKeysAccountIdParams, headers ...*common.Header) error {
 	// TODO EJR how to handle private security token
 	return client.post(ctx, nil, fmt.Sprintf("/register-participation-keys/%s", account), nil, headers)
 }
 
-func (client Client) PendingTransactionInformation(ctx context.Context, txid string, params models.GetPendingTransactionsParams, headers ...*Header) (result types.Transaction, err error) {
+func (client ClientV2) PendingTransactionInformation(ctx context.Context, txid string, params models.GetPendingTransactionsParams, headers ...*common.Header) (result types.Transaction, err error) {
 	if params.Format == "json" {
 		var response models.Transaction
 		err = client.get(ctx, &response, fmt.Sprintf("/transactions/pending/%s", txid), params, headers)
@@ -32,14 +33,14 @@ func (client Client) PendingTransactionInformation(ctx context.Context, txid str
 	return
 }
 
-func (client Client) SendRawTransaction(ctx context.Context, txBytes []byte, headers ...*Header) (txid string, err error) {
+func (client ClientV2) SendRawTransaction(ctx context.Context, txBytes []byte, headers ...*common.Header) (txid string, err error) {
 	var response models.TxId
 	err = client.post(ctx, &response, "/transactions", nil, headers)
 	txid = string(response)
 	return
 }
 
-func (client Client) PendingTransactionsByAddress(ctx context.Context, account string, params models.GetPendingTransactionsByAddressParams, headers ...*Header) (result []types.Transaction, err error) {
+func (client ClientV2) PendingTransactionsByAddress(ctx context.Context, account string, params models.GetPendingTransactionsByAddressParams, headers ...*common.Header) (result []types.Transaction, err error) {
 	err = client.get(ctx, &result, fmt.Sprintf("/accounts/%s/transactions/pending", account), params, headers)
 	if params.Format == "json" {
 		var response []models.Transaction
@@ -53,27 +54,27 @@ func (client Client) PendingTransactionsByAddress(ctx context.Context, account s
 	return
 }
 
-func (client Client) Status(ctx context.Context, headers ...*Header) (status models.NodeStatus, err error) {
+func (client ClientV2) Status(ctx context.Context, headers ...*common.Header) (status models.NodeStatus, err error) {
 	err = client.get(ctx, &status, "/status", nil, headers)
 	return
 }
 
-func (client Client) Supply(ctx context.Context, headers ...*Header) (supply models.Supply, err error) {
+func (client ClientV2) Supply(ctx context.Context, headers ...*common.Header) (supply models.Supply, err error) {
 	err = client.get(ctx, &supply, "/ledger/supply", nil, headers)
 	return
 }
 
-func (client Client) StatusAfterBlock(ctx context.Context, round uint64, headers ...*Header) (status models.NodeStatus, err error) {
+func (client ClientV2) StatusAfterBlock(ctx context.Context, round uint64, headers ...*common.Header) (status models.NodeStatus, err error) {
 	err = client.get(ctx, &status, fmt.Sprintf("/statuswait-for-block-after/%d", round), nil, headers)
 	return
 }
 
-func (client Client) AccountInformation(ctx context.Context, address string, headers ...*Header) (result models.Account, err error) {
+func (client ClientV2) AccountInformation(ctx context.Context, address string, headers ...*common.Header) (result models.Account, err error) {
 	err = client.get(ctx, &result, fmt.Sprintf("/accounts/%s", address), nil, headers)
 	return
 }
 
-func (client Client) Block(ctx context.Context, round uint64, params models.GetBlockParams, headers ...*Header) (result models.Block, err error) {
+func (client ClientV2) Block(ctx context.Context, round uint64, params models.GetBlockParams, headers ...*common.Header) (result models.Block, err error) {
 	err = client.get(ctx, &result, fmt.Sprintf("/block/%d", round), params, headers)
 	if params.Format == "json" {
 		var response models.RawBlockJson
@@ -89,7 +90,7 @@ func (client Client) Block(ctx context.Context, round uint64, params models.GetB
 	return
 }
 
-func (client Client) SuggestedParams(ctx context.Context, headers ...*Header) (params types.SuggestedParams, err error) {
+func (client ClientV2) SuggestedParams(ctx context.Context, headers ...*common.Header) (params types.SuggestedParams, err error) {
 	var response models.TransactionParams
 	err = client.get(ctx, &response, "/transactions/params", nil, headers)
 	params = types.SuggestedParams{
@@ -105,13 +106,13 @@ func (client Client) SuggestedParams(ctx context.Context, headers ...*Header) (p
 
 // Versions retrieves the VersionResponse from the running node
 // the VersionResponse includes data like version number and genesis ID
-func (client Client) Versions(ctx context.Context, headers ...*Header) (response models.Version, err error) {
+func (client ClientV2) Versions(ctx context.Context, headers ...*common.Header) (response models.Version, err error) {
 	err = client.get(ctx, &response, "/versions", nil, headers)
 	return
 }
 
 // HealthCheck does a health check on the the potentially running node,
 // returning an error if the API is down
-func (client Client) HealthCheck(ctx context.Context, headers ...*Header) error {
+func (client ClientV2) HealthCheck(ctx context.Context, headers ...*common.Header) error {
 	return client.get(ctx, nil, "/health", nil, headers)
 }
